requirements.txt
Flask==2.2.1
gunicorn==20.1.0
Werkzeug==2.2.2
pillow
